==> example/scripts/0-palette.R <==
# credit to Alec Robitaille and WEEL's study-area-figs repo for inspiration/code: https://github.com/wildlifeevoeco/study-area-figs

# Water
watercol <- '#99acc3'

# Land
montrealcol <- '#ceb99b'
canadacol <- '#ddc48d'

# Ruelles Vertes
rvcol <- '#1c3516'


# Map etc
gridcol1 <- '#323232'
gridcol2 <- '#73776F'
